services/verifier/pyproject.toml
[project]
name = "sui-human-verifier"
version = "0.1.0"
description = "FastAPI verifier service scaffold for the Sui Human MVP."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "numpy>=2.1.0,<3.0.0",
  "onnxruntime>=1.20.0,<2.0.0",
  "opencv-python>=4.10.0,<5.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "redis>=5.0.0,<6.0.0",
  "ultralytics>=8.3.0,<9.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
